+++ Sources/PrimerSDK/Classes/User Interface/TokenizationViewModels/PaymentMethodTokenizationViewModel+Logic.swift
@@ -14,7 +14,6 @@ import UIKit
 
 extension PaymentMethodTokenizationViewModel {
 
-    // TODO: FINAL_MIGRATION
     @objc
     func start() {
         firstly {
@@ -58,15 +57,63 @@ extension PaymentMethodTokenizationViewModel {
                                                                  diagnosticsId: UUID().uuidString)
                     }
 
-                    self.showResultScreenIfNeeded(error: primerErr)
+                    DispatchQueue.main.async {
+                        self.showResultScreenIfNeeded(error: primerErr)
+                    }
                     return PrimerDelegateProxy.raisePrimerDidFailWithError(primerErr, data: self.paymentCheckoutData)
                 }
                 .done { merchantErrorMessage in
-                    self.handleFailureFlow(errorMessage: merchantErrorMessage)
+                    DispatchQueue.main.async {
+                        self.handleFailureFlow(errorMessage: merchantErrorMessage)
+                    }
                 }
                 // The above promises will never end up on error.
                 .catch { _ in
                     self.logger.error(message: "Unselection of payment method failed - this should never happen ...")
+                }
+            }
+        }
+    }
+
+    @objc
+    func start_async() {
+        Task {
+            do {
+                self.paymentMethodTokenData = try await startTokenizationFlow()
+                processPaymentMethodTokenData()
+                await uiManager.primerRootViewController?.enableUserInteraction(true)
+            } catch {
+                await uiManager.primerRootViewController?.enableUserInteraction(true)
+                let clientSessionActionsModule: ClientSessionActionsProtocol = ClientSessionActionsModule()
+
+                if let primerErr = error as? PrimerError,
+                   case .cancelled = primerErr,
+                   PrimerInternal.shared.sdkIntegrationType == .dropIn,
+                   self.config.type == PrimerPaymentMethodType.applePay.rawValue ||
+                   self.config.type == PrimerPaymentMethodType.adyenIDeal.rawValue ||
+                   self.config.type == PrimerPaymentMethodType.payPal.rawValue {
+                    try? await clientSessionActionsModule.unselectPaymentMethodIfNeeded()
+                    await PrimerUIManager.primerRootViewController?.popToMainScreen(completion: nil)
+
+                } else {
+                    do {
+                        try await clientSessionActionsModule.unselectPaymentMethodIfNeeded()
+
+                        let primerErr: PrimerError
+                        if let error = error as? PrimerError {
+                            primerErr = error
+                        } else {
+                            primerErr = PrimerError.underlyingErrors(errors: [error],
+                                                                     userInfo: .errorUserInfoDictionary(),
+                                                                     diagnosticsId: UUID().uuidString)
+                        }
+                        await showResultScreenIfNeeded(error: primerErr)
+
+                        let merchantErrorMessage = await PrimerDelegateProxy.raisePrimerDidFailWithError(primerErr, data: paymentCheckoutData)
+                        await handleFailureFlow(errorMessage: merchantErrorMessage)
+                    } catch {
+                        logger.error(message: "Unselection of payment method failed - this should never happen ...")
+                    }
                 }
             }
         }
@@ -86,7 +133,9 @@ extension PaymentMethodTokenizationViewModel {
 
     func processVaultPaymentMethodTokenData() {
         PrimerDelegateProxy.primerDidTokenizePaymentMethod(self.paymentMethodTokenData!) { _ in }
-        self.handleSuccessfulFlow()
+        DispatchQueue.main.async {
+            self.handleSuccessfulFlow()
+        }
     }
 
     func processCheckoutPaymentMethodTokenData() {
@@ -110,8 +159,10 @@ extension PaymentMethodTokenizationViewModel {
                 PrimerDelegateProxy.primerDidCompleteCheckoutWithData(checkoutData)
             }
 
-            self.showResultScreenIfNeeded()
-            self.handleSuccessfulFlow()
+            DispatchQueue.main.async {
+                self.showResultScreenIfNeeded()
+                self.handleSuccessfulFlow()
+            }
         }
         .ensure {
             PrimerUIManager.primerRootViewController?.enableUserInteraction(true)
@@ -153,14 +204,90 @@ extension PaymentMethodTokenizationViewModel {
                                                                  diagnosticsId: UUID().uuidString)
                     }
                     self.setCheckoutDataFromError(primerErr)
-                    self.showResultScreenIfNeeded(error: primerErr)
+                    DispatchQueue.main.async {
+                        self.showResultScreenIfNeeded(error: primerErr)
+                    }
                     return PrimerDelegateProxy.raisePrimerDidFailWithError(primerErr, data: self.paymentCheckoutData)
                 }
                 .done { merchantErrorMessage in
-                    self.handleFailureFlow(errorMessage: merchantErrorMessage)
+                    DispatchQueue.main.async {
+                        self.handleFailureFlow(errorMessage: merchantErrorMessage)
+                    }
                 }
                 // The above promises will never end up on error.
                 .catch { _ in }
+            }
+        }
+    }
+
+    func processCheckoutPaymentMethodTokenData_async() {
+        didStartPayment?()
+        didStartPayment = nil
+
+        if config.internalPaymentMethodType != .klarna {
+            PrimerUIManager.primerRootViewController?.showLoadingScreenIfNeeded(
+                imageView: uiModule.makeIconImageView(withDimension: 24.0),
+                message: nil
+            )
+        }
+
+        Task {
+            defer {
+                DispatchQueue.main.async {
+                    PrimerUIManager.primerRootViewController?.enableUserInteraction(true)
+                }
+            }
+            do {
+                guard let paymentMethodTokenData else {
+                    throw PrimerError.invalidValue(
+                        key: "paymentMethodTokenData",
+                        value: "Payment method token data is not valid",
+                        userInfo: .errorUserInfoDictionary(),
+                        diagnosticsId: UUID().uuidString
+                    )
+                }
+                let checkoutData = try await startPaymentFlow(withPaymentMethodTokenData: paymentMethodTokenData)
+
+                didFinishPayment?(nil)
+                nullifyEventCallbacks()
+
+                if PrimerSettings.current.paymentHandling == .auto, let checkoutData {
+                    await PrimerDelegateProxy.primerDidCompleteCheckoutWithData(checkoutData)
+                }
+
+                await showResultScreenIfNeeded()
+                await handleSuccessfulFlow()
+            } catch {
+                didFinishPayment?(error)
+                nullifyEventCallbacks()
+
+                let clientSessionActionsModule: ClientSessionActionsProtocol = ClientSessionActionsModule()
+
+                if let primerErr = error as? PrimerError,
+                   case .cancelled = primerErr,
+                   PrimerInternal.shared.sdkIntegrationType == .dropIn,
+                   PrimerInternal.shared.selectedPaymentMethodType == nil,
+                   self.config.implementationType == .webRedirect ||
+                   self.config.type == PrimerPaymentMethodType.applePay.rawValue ||
+                   self.config.type == PrimerPaymentMethodType.adyenIDeal.rawValue ||
+                   self.config.type == PrimerPaymentMethodType.payPal.rawValue {
+                    try? await clientSessionActionsModule.unselectPaymentMethodIfNeeded()
+                } else {
+                    try? await clientSessionActionsModule.unselectPaymentMethodIfNeeded()
+
+                    let primerErr: PrimerError
+                    if let error = error as? PrimerError {
+                        primerErr = error
+                    } else {
+                        primerErr = PrimerError.underlyingErrors(errors: [error],
+                                                                 userInfo: .errorUserInfoDictionary(),
+                                                                 diagnosticsId: UUID().uuidString)
+                    }
+                    setCheckoutDataFromError(primerErr)
+                    await showResultScreenIfNeeded(error: primerErr)
+                    let merchantErrorMessage = await PrimerDelegateProxy.raisePrimerDidFailWithError(primerErr, data: paymentCheckoutData)
+                    await handleFailureFlow(errorMessage: merchantErrorMessage)
+                }
             }
         }
     }
@@ -264,9 +391,9 @@ extension PaymentMethodTokenizationViewModel {
                     }
                 }
 
-                return self.paymentCheckoutData
+                return paymentCheckoutData
             } catch is CancellationError {
-                let cancelledError = PrimerError.cancelled(paymentMethodType: self.config.type,
+                let cancelledError = PrimerError.cancelled(paymentMethodType: config.type,
                                                            userInfo: .errorUserInfoDictionary(),
                                                            diagnosticsId: UUID().uuidString)
                 ErrorHandler.handle(error: cancelledError)
@@ -422,17 +549,16 @@ extension PaymentMethodTokenizationViewModel {
         withPaymentMethodTokenData paymentMethodTokenData: PrimerPaymentMethodTokenData
     ) async throws -> DecodedJWTToken? {
         if PrimerSettings.current.paymentHandling == .manual {
-            return try await startManualPaymentFlowAndFetchToken(withPaymentMethodTokenData: paymentMethodTokenData)
+            try await startManualPaymentFlowAndFetchToken(withPaymentMethodTokenData: paymentMethodTokenData)
         } else {
-            return try await startAutomaticPaymentFlowAndFetchToken(withPaymentMethodTokenData: paymentMethodTokenData)
+            try await startAutomaticPaymentFlowAndFetchToken(withPaymentMethodTokenData: paymentMethodTokenData)
         }
     }
 
     func startManualPaymentFlowAndFetchToken(
         withPaymentMethodTokenData paymentMethodTokenData: PrimerPaymentMethodTokenData
     ) async throws -> DecodedJWTToken? {
-        let resumeDecision = try await PrimerDelegateProxy.primerDidTokenizePaymentMethod(paymentMethodTokenData)
-
+        let resumeDecision = await PrimerDelegateProxy.primerDidTokenizePaymentMethod(paymentMethodTokenData)
         if let resumeDecisionType = resumeDecision.type as? PrimerResumeDecision.DecisionType {
             switch resumeDecisionType {
             case .succeed:
@@ -453,7 +579,7 @@ extension PaymentMethodTokenizationViewModel {
                 return decodedJWTToken
 
             case .fail(let message):
-                var merchantErr: Error!
+                let merchantErr: Error
                 if let message = message {
                     let err = PrimerError.merchantError(message: message,
                                                         userInfo: .errorUserInfoDictionary(),
@@ -464,7 +590,6 @@ extension PaymentMethodTokenizationViewModel {
                 }
                 throw merchantErr
             }
-
         } else if let resumeDecisionType = resumeDecision.type as? PrimerHeadlessUniversalCheckoutResumeDecision.DecisionType {
             switch resumeDecisionType {
             case .continueWithNewClientToken(let newClientToken):
@@ -590,19 +715,19 @@ extension PaymentMethodTokenizationViewModel {
 
     func handleResumeStepsBasedOnSDKSettings(resumeToken: String) async throws -> PrimerCheckoutData? {
         if PrimerSettings.current.paymentHandling == .manual {
-            return try await handleManualResumeStepsBasedOnSDKSettings(resumeToken: resumeToken)
+            try await handleManualResumeStepsBasedOnSDKSettings(resumeToken: resumeToken)
         } else {
-            return try await handleAutomaticResumeStepsBasedOnSDKSettings(resumeToken: resumeToken)
+            try await handleAutomaticResumeStepsBasedOnSDKSettings(resumeToken: resumeToken)
         }
     }
 
     private func handleManualResumeStepsBasedOnSDKSettings(resumeToken: String) async throws -> PrimerCheckoutData? {
-        let resumeDecision = try await PrimerDelegateProxy.primerDidResumeWith(resumeToken)
+        let resumeDecision = await PrimerDelegateProxy.primerDidResumeWith(resumeToken)
 
         if let resumeDecisionType = resumeDecision.type as? PrimerResumeDecision.DecisionType {
             switch resumeDecisionType {
             case .fail(let message):
-                var merchantErr: Error!
+                let merchantErr: Error
                 if let message {
                     let err = PrimerError.merchantError(message: message,
                                                         userInfo: .errorUserInfoDictionary(),
@@ -616,7 +741,6 @@ extension PaymentMethodTokenizationViewModel {
             case .succeed, .continueWithNewClientToken:
                 return nil
             }
-
         } else if resumeDecision.type is PrimerHeadlessUniversalCheckoutResumeDecision.DecisionType {
             return nil
         } else {
@@ -641,6 +765,7 @@ extension PaymentMethodTokenizationViewModel {
 
     // This method will show the new design for result screen with a specific state: e.g. Error state or Success state
     // For now we will use it only for STRIPE_ACH implementation
+    @MainActor
     func showResultScreenIfNeeded(error: PrimerError? = nil) {
         guard let paymentMethodType = config.internalPaymentMethodType,
               paymentMethodType == .stripeAch else {
@@ -649,9 +774,10 @@ extension PaymentMethodTokenizationViewModel {
         PrimerUIManager.showResultScreen(for: paymentMethodType, error: error)
     }
 
+    @MainActor
     func handleFailureFlow(errorMessage: String?) {
         if config.internalPaymentMethodType != .stripeAch {
-            let categories = self.config.paymentMethodManagerCategories
+            let categories = config.paymentMethodManagerCategories
             PrimerUIManager.dismissOrShowResultScreen(
                 type: .failure,
                 paymentMethodManagerCategories: categories ?? [],
@@ -720,7 +846,7 @@ Make sure you call the decision handler otherwise the SDK will hang.
             }
         }
 
-        let paymentCreationDecision = try await PrimerDelegateProxy.primerWillCreatePaymentWithData(checkoutPaymentMethodData)
+        let paymentCreationDecision = await PrimerDelegateProxy.primerWillCreatePaymentWithData(checkoutPaymentMethodData)
         decisionHandlerHasBeenCalled = true
 
         switch paymentCreationDecision.type {
@@ -729,7 +855,6 @@ Make sure you call the decision handler otherwise the SDK will hang.
                                                   userInfo: .errorUserInfoDictionary(),
                                                   diagnosticsId: UUID().uuidString)
             throw error
-
         case .continue:
             return
         }