+++ Tests/Utilities/Mocks/MockPaymentMethodTokenizationViewModel.swift
@@ -73,8 +73,9 @@ class MockPaymentMethodTokenizationViewModel: NSObject, PaymentMethodTokenizatio
             self.paymentMethodTokenData = paymentMethodTokenData
 
             if PrimerInternal.shared.intent == .vault {
-                self.handleSuccessfulFlow()
-
+                DispatchQueue.main.async{
+                    self.handleSuccessfulFlow()
+                }
             } else {
                 self.didStartPayment?()
                 self.didStartPayment = nil
@@ -85,7 +86,9 @@ class MockPaymentMethodTokenizationViewModel: NSObject, PaymentMethodTokenizatio
                 .done { _ in
                     self.didFinishPayment?(nil)
                     self.nullifyEventCallbacks()
-                    self.handleSuccessfulFlow()
+                    DispatchQueue.main.async {
+                        self.handleSuccessfulFlow()
+                    }
                 }
                 .catch { err in
                     self.didFinishPayment?(err)
@@ -176,7 +179,7 @@ class MockPaymentMethodTokenizationViewModel: NSObject, PaymentMethodTokenizatio
     }
 
     func performTokenizationStep() async throws {
-        PrimerDelegateProxy.primerHeadlessUniversalCheckoutDidStartTokenization(for: config.type)
+        await PrimerDelegateProxy.primerHeadlessUniversalCheckoutDidStartTokenization(for: config.type)
         try await checkoutEventsNotifierModule.fireDidStartTokenizationEvent()
         let paymentMethodTokenData = try await tokenize()
         self.paymentMethodTokenData = paymentMethodTokenData
@@ -304,8 +307,10 @@ class MockPaymentMethodTokenizationViewModel: NSObject, PaymentMethodTokenizatio
         }
     }
 
+    @MainActor
     func handleSuccessfulFlow() {}
 
+    @MainActor
     func handleFailureFlow(errorMessage: String?) {}
 
     func submitButtonTapped() {}